requirements.txt
opencv-python==4.8.1.78
scenedetect[opencv]==0.6.2
numpy==1.24.3
moviepy==1.0.3
ffmpeg-python==0.2.0
openai-whisper==20230918
tqdm==4.66.1
Pillow==10.1.0
scipy==1.11.4
librosa==0.10.1
torch>=2.0.0
torchvision>=0.15.0